reformatted

+++ src/main/java/vlaship/backoffice/controller/AbstractController.java
@@ -14,31 +14,31 @@ import java.util.List;
 @SecurityRequirement(name = "Bearer Authentication")
 public class AbstractController<M extends Model, D extends Dto> {
 
-	private final Facade<M, D> facade;
+    private final Facade<M, D> facade;
 
-	@PutMapping("/update")
-	public ResponseEntity<D> update(@Valid final @RequestBody D dto) {
-		return ResponseEntity.accepted().body(facade.update(dto));
-	}
+    @PutMapping("/update")
+    public ResponseEntity<D> update(@Valid final @RequestBody D dto) {
+        return ResponseEntity.accepted().body(facade.update(dto));
+    }
 
-	@DeleteMapping("/delete/{id}")
-	public ResponseEntity<Void> delete(final @PathVariable("id") Integer id) {
-		facade.delete(id);
-		return ResponseEntity.accepted().build();
-	}
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<Void> delete(final @PathVariable("id") Long id) {
+        facade.delete(id);
+        return ResponseEntity.accepted().build();
+    }
 
-	@GetMapping("/{id}")
-	public ResponseEntity<D> find(final @PathVariable("id") Integer id) {
-		return ResponseEntity.ok(facade.find(id));
-	}
+    @GetMapping("/{id}")
+    public ResponseEntity<D> find(final @PathVariable("id") Long id) {
+        return ResponseEntity.ok(facade.find(id));
+    }
 
-	@GetMapping("/list")
-	public ResponseEntity<List<D>> findAll(final Pageable pageable) {
-		return ResponseEntity.ok(facade.findAll(pageable));
-	}
+    @GetMapping("/list")
+    public ResponseEntity<List<D>> findAll(final Pageable pageable) {
+        return ResponseEntity.ok(facade.findAll(pageable));
+    }
 
-	public AbstractController(final Facade<M, D> facade) {
-		this.facade = facade;
-	}
+    public AbstractController(final Facade<M, D> facade) {
+        this.facade = facade;
+    }
 
 }
