+++ src/main/java/vlaship/backoffice/exception/BadRequestException.java
@@ -2,7 +2,7 @@ package vlaship.backoffice.exception;
 
 public class BadRequestException extends AbstractException {
 
-    public BadRequestException(final String model, final int id) {
+    public BadRequestException(final String model, final long id) {
         super("Can't find " + model + " with ID = " + id);
     }
