+++ src/main/java/vlaship/backoffice/exception/NotFoundException.java
@@ -2,8 +2,8 @@ package vlaship.backoffice.exception;
 
 public class NotFoundException extends AbstractException {
 
-	public NotFoundException(final String model, final int id) {
-		super("Can't find " + model + " with ID = " + id);
-	}
+    public NotFoundException(final String model, final long id) {
+        super("Can't find " + model + " with ID = " + id);
+    }
 
 }
