+++ src/main/java/vlaship/backoffice/model/Model.java
@@ -4,7 +4,7 @@ import org.springframework.lang.NonNull;
 
 public interface Model {
 
-	@NonNull
-	Integer getId();
+    @NonNull
+    Long getId();
 
 }
