+++ src/main/java/vlaship/backoffice/repository/PriceRepository.java
@@ -9,14 +9,14 @@ import java.math.BigDecimal;
 import java.util.Currency;
 import java.util.List;
 
-public interface PriceRepository extends JpaRepository<Price, Integer> {
+public interface PriceRepository extends JpaRepository<Price, Long> {
 
-	List<Price> findAllByAmountBetweenAndCurrency(BigDecimal from, BigDecimal to, Currency currency, Pageable pageable);
+    List<Price> findAllByAmountBetweenAndCurrency(BigDecimal from, BigDecimal to, Currency currency, Pageable pageable);
 
-	List<Price> findAllByCurrency(Currency currency, Pageable pageable);
+    List<Price> findAllByCurrency(Currency currency, Pageable pageable);
 
-	List<Price> findAllByProduct(Product product, Pageable pageable);
+    List<Price> findAllByProduct(Product product, Pageable pageable);
 
-	Integer countAllByProduct(Product product);
+    Integer countAllByProduct(Product product);
 
 }
