+++ src/main/java/vlaship/backoffice/repository/UserRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
-	Optional<User> findByName(String name);
+    Optional<User> findByName(String name);
 
 }
